backend/requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==2.1.0
fastapi==0.104.0
greenlet==2.0.2
h11==0.13.0
httpx==0.25.0
idna==3.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.7
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomli==2.0.1
urllib3==1.26.12
uvicorn==0.18.2
zipp==3.8.1
